+++ public_html/js/draw_token.js
@@ -1,27 +1,27 @@
 // draw_token.js
 // (c) inoro 2019 GPL v3
 
-function draw_black_token(ctx, cx, cy, m, tBw) {
+function draw_top_token(ctx, cx, cy, m, tBw) {
     ctx.beginPath();
     ctx.arc(cx*m, cy*m, 10*m-(tBw/2), 0, 2*Math.PI);
     ctx.closePath();
-    ctx.strokeStyle = window.bgjs.black_stroke_color;
+    ctx.strokeStyle = window.bgjs.top_stroke_color;
     ctx.stroke();
-    ctx.fillStyle = window.bgjs.black_fill_color;
+    ctx.fillStyle = window.bgjs.top_fill_color;
     ctx.fill();
 }
 
-function draw_white_token(ctx, cx, cy, m, tBw) {
+function draw_bottom_token(ctx, cx, cy, m, tBw) {
     ctx.beginPath();
     ctx.arc(cx*m, cy*m, 10*m-(tBw/2), 0, 2*Math.PI);
     ctx.closePath();
-    ctx.strokeStyle = window.bgjs.white_stroke_color;
+    ctx.strokeStyle = window.bgjs.bottom_stroke_color;
     ctx.stroke();
-    ctx.fillStyle = window.bgjs.white_fill_color;
+    ctx.fillStyle = window.bgjs.bottom_fill_color;
     ctx.fill();
 }
 
-function draw_black_token_taked(ctx, m, tBw, num) {
+function draw_top_token_taked(ctx, m, tBw, num) {
     var
         tx = xyT(7),
         ty = xyT(5.5);
@@ -29,18 +29,18 @@ function draw_black_token_taked(ctx, m, tBw, num) {
     ctx.beginPath();
     ctx.arc(tx*m, ty*m, 10*m-(tBw/2), 0, 2*Math.PI);
     ctx.closePath();
-    ctx.strokeStyle = window.bgjs.black_stroke_color;
+    ctx.strokeStyle = window.bgjs.top_stroke_color;
     ctx.stroke();
-    ctx.fillStyle = window.bgjs.black_fill_color;
+    ctx.fillStyle = window.bgjs.top_fill_color;
     ctx.fill();
 
     if (num > 1) {
-        ctx.fillStyle = window.bgjs.black_text_color;
+        ctx.fillStyle = window.bgjs.top_text_color;
         ctx.fillText(num, tx*m, (ty+3.2)*m);
     }
 }
 
-function draw_white_token_taked(ctx, m, tBw, num) {
+function draw_bottom_token_taked(ctx, m, tBw, num) {
     var
         tx = xyT(7),
         ty = xyT(7.5);
@@ -48,13 +48,13 @@ function draw_white_token_taked(ctx, m, tBw, num) {
     ctx.beginPath();
     ctx.arc(tx*m, ty*m, 10*m-(tBw/2), 0, 2*Math.PI);
     ctx.closePath();
-    ctx.strokeStyle = window.bgjs.white_stroke_color;
+    ctx.strokeStyle = window.bgjs.bottom_stroke_color;
     ctx.stroke();
-    ctx.fillStyle = window.bgjs.white_fill_color;
+    ctx.fillStyle = window.bgjs.bottom_fill_color;
     ctx.fill();
 
     if (num > 1) {
-        ctx.fillStyle = window.bgjs.white_text_color;
+        ctx.fillStyle = window.bgjs.bottom_text_color;
         ctx.fillText(num, tx*m, (ty+3.2)*m);
     }
 }
@@ -81,12 +81,12 @@ function draw_layout(ctx, m, tkstate) {
                 overflow = true;
                 break;
             }
-            draw_black_token(ctx, xyT(vboard[i][e][0]), xyT(vboard[i][e][1]), m, tBw);
+            draw_top_token(ctx, xyT(vboard[i][e][0]), xyT(vboard[i][e][1]), m, tBw);
             e++;
         }
         if (overflow) {
             last_e = 4;
-            ctx.fillStyle = window.bgjs.black_text_color;
+            ctx.fillStyle = window.bgjs.top_text_color;
             ctx.fillText(tkstate[i][0], xyT(vboard[i][last_e][0])*m, (xyT(vboard[i][last_e][1])+3.2)*m);
         }
 
@@ -98,23 +98,23 @@ function draw_layout(ctx, m, tkstate) {
                 overflow = true;
                 break;
             }
-            draw_white_token(ctx, xyT(vboard[i][e][0]), xyT(vboard[i][e][1]), m, tBw);
+            draw_bottom_token(ctx, xyT(vboard[i][e][0]), xyT(vboard[i][e][1]), m, tBw);
             e++;
         }
         if (overflow) {
             last_e = 4;
-            ctx.fillStyle = window.bgjs.white_text_color;
+            ctx.fillStyle = window.bgjs.bottom_text_color;
             ctx.fillText(tkstate[i][1], xyT(vboard[i][last_e][0])*m, (xyT(vboard[i][last_e][1])+3.2)*m);
         }
 
         i++;
     }
 
-    if (window.bgjs.black_tokens_taked > 0) {
-        draw_black_token_taked(ctx, m, tBw, window.bgjs.black_tokens_taked);
+    if (window.bgjs.top_tokens_taked > 0) {
+        draw_top_token_taked(ctx, m, tBw, window.bgjs.top_tokens_taked);
     }
-    if (window.bgjs.white_tokens_taked > 0) {
-        draw_white_token_taked(ctx, m, tBw, window.bgjs.white_tokens_taked);
+    if (window.bgjs.bottom_tokens_taked > 0) {
+        draw_bottom_token_taked(ctx, m, tBw, window.bgjs.bottom_tokens_taked);
     }
 }