+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [20190913b] - 2019-09-13 20:15
 ### Changed
 - Ahora se puede seleccionar visualmente las fichas comidas de ambos jugadores.
+- Todas las variables "black" ahora se llaman "top".
+- Todas las variables "white" ahora se llaman "bottom".
+- Todas las variables "rojo" ahora se llaman "top_color".
 
 ## [20190913a] - 2019-09-13 11:15
 ### Added